Extract variable::Value to separate module
Making the variable module more compact.
