+++ yash-env/src/variable.rs
@@ -46,10 +46,8 @@
 
 #[cfg(doc)]
 use crate::Env;
-use either::{Left, Right};
 use itertools::Itertools;
 use std::borrow::Borrow;
-use std::borrow::Cow;
 use std::collections::HashMap;
 use std::ffi::CString;
 use std::fmt::Write;
@@ -58,125 +56,9 @@ use std::iter::FusedIterator;
 use thiserror::Error;
 use yash_syntax::source::Location;
 
-/// Value of a variable.
-#[derive(Clone, Debug, Eq, PartialEq)]
-pub enum Value {
-    /// Single string.
-    Scalar(String),
-    /// Array of strings.
-    Array(Vec<String>),
-}
+mod value;
 
-pub use Value::*;
-
-impl Value {
-    /// Creates a scalar value.
-    #[must_use]
-    pub fn scalar<S: Into<String>>(value: S) -> Self {
-        Scalar(value.into())
-    }
-
-    /// Creates an array value.
-    #[must_use]
-    pub fn array<I, S>(values: I) -> Self
-    where
-        I: IntoIterator<Item = S>,
-        S: Into<String>,
-    {
-        Array(values.into_iter().map(Into::into).collect())
-    }
-
-    /// Splits the value by colons.
-    ///
-    /// If this value is `Scalar`, the value is separated at each occurrence of
-    /// colon (`:`). For `Array`, each array item is returned without further
-    /// splitting the value.
-    ///
-    /// ```
-    /// # use yash_env::variable::Value;
-    /// let scalar = Value::scalar("/usr/local/bin:/usr/bin:/bin");
-    /// let values: Vec<&str> = scalar.split().collect();
-    /// assert_eq!(values, ["/usr/local/bin", "/usr/bin", "/bin"]);
-    /// ```
-    ///
-    /// ```
-    /// # use yash_env::variable::Value;
-    /// let array = Value::array(vec!["foo", "bar"]);
-    /// let values: Vec<&str> = array.split().collect();
-    /// assert_eq!(values, ["foo", "bar"]);
-    /// ```
-    pub fn split(&self) -> impl Iterator<Item = &str> {
-        match self {
-            Scalar(value) => Left(value.split(':')),
-            Array(values) => Right(values.iter().map(String::as_str)),
-        }
-    }
-
-    /// Quotes the value in a format suitable for re-parsing.
-    ///
-    /// This function returns a temporary wrapper of `self`. To obtain a string
-    /// representation of the quoted value, you can use the `Display` or
-    /// `Into<Cow<str>>` implementation for the returned object.
-    ///
-    /// See [`yash_quote`] for details of quoting.
-    ///
-    /// ```
-    /// # use yash_env::variable::Value;
-    /// let scalar = Value::scalar("foo bar");
-    /// assert_eq!(scalar.quote().to_string(), "'foo bar'");
-    /// let array = Value::array(vec!["1", "", "'\\'"]);
-    /// assert_eq!(array.quote().to_string(), r#"(1 '' "'\\'")"#);
-    /// ```
-    pub fn quote(&self) -> QuotedValue {
-        QuotedValue::from(self)
-    }
-}
-
-/// Wrapper of [`Value`] for [quoting](Value::quote).
-#[derive(Clone, Copy, Debug)]
-pub struct QuotedValue<'a> {
-    value: &'a Value,
-}
-
-/// Writes a quoted version of the value to the formatter.
-impl<'a> std::fmt::Display for QuotedValue<'a> {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self.value {
-            Scalar(value) => yash_quote::quoted(value).fmt(f),
-            Array(values) => write!(
-                f,
-                "({})",
-                values
-                    .iter()
-                    .format_with(" ", |value, f| f(&yash_quote::quoted(value)))
-            ),
-        }
-    }
-}
-
-/// Wraps a value in `QuotedValue`.
-impl<'a> From<&'a Value> for QuotedValue<'a> {
-    fn from(value: &'a Value) -> Self {
-        QuotedValue { value }
-    }
-}
-
-/// Extracts the wrapped reference to the value.
-impl AsRef<Value> for QuotedValue<'_> {
-    fn as_ref(&self) -> &Value {
-        self.value
-    }
-}
-
-/// Constructs a quoted string.
-impl<'a> From<QuotedValue<'a>> for Cow<'a, str> {
-    fn from(value: QuotedValue<'a>) -> Self {
-        match value.value {
-            Scalar(value) => yash_quote::quote(value),
-            Array(_values) => value.to_string().into(),
-        }
-    }
-}
+pub use self::value::{Array, Scalar, Value};
 
 mod quirk;
